Refactor and fix with HY's suggestions.
